Fix type

+++ trunk/examples/programs/toy/showcase/2001TACAS-BallPodelskiRajamani.c
@@ -6,7 +6,7 @@
  * http://link.springer.com/chapter/10.1007%2F3-540-45319-9_19
  */
 
-extern int __VERIFIER_nondet_short(void);
+extern short __VERIFIER_nondet_short(void);
 
 int main() {
     int x, y, z, w;
